# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sutradhar API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/categories": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Categories",
        "description": "Latest version of every category. Public.",
        "operationId": "list_categories_api_v1_categories_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories/{slug}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Read Category",
        "description": "Latest active version of one category.",
        "operationId": "read_category_api_v1_categories__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories/{slug}/versions": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Read Versions",
        "description": "Every published version, oldest first.",
        "operationId": "read_versions_api_v1_categories__slug__versions_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryVersionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories/{slug}/v/{version}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Read Pinned Version",
        "description": "One pinned version, retired or not.\n\nRetired versions still resolve: existing items reference them, and a\nverification that 404s because a category was retired would be a broken\nprovenance record.",
        "operationId": "read_pinned_version_api_v1_categories__slug__v__version__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories/{slug}/validate": {
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Validate Attributes",
        "description": "Dry-run an attribute payload against a category. Writes nothing.\n\nLets a client check a form before submitting it, and gives the live-demo\nsequence a real request that proves a just-created category is usable.\nItem registration itself is Phase 6.\n\nRate limited, unlike the reads beside it. This is the only unauthenticated\nendpoint in the system that does real work on a caller-supplied body: it\nloads a schema and runs a validator over arbitrary JSON. The reads answer\nfrom a cached registry and cost a lookup; this costs compilation, and on a\nsingle free-tier instance an unmetered one of those is a way to spend the\nprocess's CPU from outside with no account and no cost to the caller.",
        "operationId": "validate_attributes_api_v1_categories__slug__validate_post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateAttributesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateAttributesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/categories": {
      "post": {
        "tags": [
          "catalog-admin"
        ],
        "summary": "Create Category",
        "description": "Create a category at v1, usable immediately.\n\nAccepts a category JSON document as the request body, so publishing one on\nstage is ``curl -d @banarasi.json`` and nothing else.",
        "operationId": "create_category_api_v1_admin_categories_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/categories/{slug}/versions": {
      "post": {
        "tags": [
          "catalog-admin"
        ],
        "summary": "Create Version",
        "description": "Publish v(n+1), and report what it changes.\n\nThe diff is returned so the operator sees the consequence at the moment\nthey cause it, rather than discovering it when registrations start failing.\nExisting items are unaffected either way -- they are pinned to their own\nversion.",
        "operationId": "create_version_api_v1_admin_categories__slug__versions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVersionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/categories/{slug}": {
      "patch": {
        "tags": [
          "catalog-admin"
        ],
        "summary": "Update Category",
        "description": "Update display name and active flag. Slug and schema are unreachable.\n\nNeither field exists on ``UpdateCategoryRequest``, so no request shape can\nchange them -- a slug is referenced by URLs and printed tags, and a schema\nchange is a new version by definition.",
        "operationId": "update_category_api_v1_admin_categories__slug__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "is_textile": {
            "type": "boolean",
            "title": "Is Textile"
          },
          "quantity_unit": {
            "type": "string",
            "title": "Quantity Unit"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "attribute_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attribute Schema"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "display_name",
          "is_textile",
          "quantity_unit",
          "schema_version",
          "is_active",
          "created_at",
          "attribute_schema"
        ],
        "title": "CategoryDetail",
        "description": "A category with its schema."
      },
      "CategoryListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CategorySummary"
            },
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "additionalProperties": true,
            "type": "object",
            "title": "Pagination"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "CategoryListResponse",
        "description": "Collection envelope, per app.core.envelope."
      },
      "CategorySummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "is_textile": {
            "type": "boolean",
            "title": "Is Textile"
          },
          "quantity_unit": {
            "type": "string",
            "title": "Quantity Unit"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "display_name",
          "is_textile",
          "quantity_unit",
          "schema_version",
          "is_active",
          "created_at"
        ],
        "title": "CategorySummary",
        "description": "A category in a listing."
      },
      "CategoryVersionsResponse": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/CategorySummary"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "data"
        ],
        "title": "CategoryVersionsResponse"
      },
...
      "CreateCategoryRequest": {
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Slug"
          },
          "display_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name"
          },
          "is_textile": {
            "type": "boolean",
            "title": "Is Textile",
            "default": true
          },
          "quantity_unit": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Quantity Unit"
          },
          "attribute_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attribute Schema"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "display_name",
          "quantity_unit",
          "attribute_schema"
        ],
        "title": "CreateCategoryRequest",
        "description": "Create a category at v1.\n\n``schema_version`` is absent on purpose: v1 is implied, and letting a caller\npick a starting version would allow gaps that make \"the previous version\"\nambiguous."
      },
      "CreateVersionRequest": {
        "properties": {
          "attribute_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attribute Schema"
          }
        },
        "type": "object",
        "required": [
          "attribute_schema"
        ],
        "title": "CreateVersionRequest",
        "description": "Publish v(n+1) of an existing category. Only the schema changes."
      },
      "CreateVersionResponse": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/CategoryDetail"
          },
          "diff": {
            "$ref": "#/components/schemas/SchemaDiff"
          },
          "breaking": {
            "type": "boolean",
            "title": "Breaking"
          }
        },
        "type": "object",
        "required": [
          "category",
          "diff",
          "breaking"
        ],
        "title": "CreateVersionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SchemaDiff": {
        "properties": {
          "added": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Added"
          },
          "removed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Removed"
          },
          "type_changed": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Type Changed"
          },
          "newly_required": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Newly Required"
          },
          "no_longer_required": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "No Longer Required"
          }
        },
        "type": "object",
        "required": [
          "added",
          "removed",
          "type_changed",
          "newly_required",
          "no_longer_required"
        ],
        "title": "SchemaDiff",
        "description": "What publishing this version changes, relative to the previous one.\n\nReturned from the create-version endpoint so an operator sees the\nconsequence at the moment they cause it, rather than discovering it when a\nweaver's next registration starts failing."
      },
...
      "UpdateCategoryRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "UpdateCategoryRequest",
        "description": "Mutable category metadata.\n\n``slug`` and ``attribute_schema`` are absent, not optional. A slug is\nreferenced by URLs and printed tags; a schema change is a new version by\ndefinition. Neither is reachable through this model, so neither can be\nchanged by any shape of request body."
      },
...
      "ValidateAttributesRequest": {
        "properties": {
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          }
        },
        "type": "object",
        "required": [
          "attributes"
        ],
        "title": "ValidateAttributesRequest",
        "description": "Dry-run attribute check. Writes nothing."
      },
      "ValidateAttributesResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          }
        },
        "type": "object",
        "required": [
          "valid",
          "slug",
          "schema_version"
        ],
        "title": "ValidateAttributesResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}